General code cleanup as well as changing the AccountID properties to utilize longs instead of ints (Int32 vs Int64).
The usage of ints (Int32) was what was causing the unhandled Newtonsoft.JsonReaderException.

We may also want to consider updating Newtonsoft to 12.0.3 across each of the different projects.

+++ API Models/LolItemSetsItemSets.cs
@@ -2,7 +2,7 @@ namespace LCU.NET
 {
     public class LolItemSetsItemSets
     {
-        public int accountId { get; set; }
+        public long accountId { get; set; }
         public LolItemSetsItemSet[] itemSets { get; set; }
         public long timestamp { get; set; }
     }
